BeaconSearchResults.js -> .tsx
